FIX(agent): Implemented better structure and updated README and gitignore and venv setup
