+++ .env.example
@@ -3,4 +3,4 @@ PRIVATE_KEY=
 SOURCE_ADDRESS=
 TARGET_ADDRESS=
 CONTRACT_ADDRESS=
-CHAIN_ID=17002
+CHAIN_ID=11155107
