+++ setup.sh
@@ -3,8 +3,11 @@
 # Install required Python packages
 pip3 install -r requirements.txt
 
+#activate venv
+source venv/bin/activate
+
 # Run the agent script
-python3 autonomous_agent.py
+python3 main.py
 
 # Run the unit test script
 python3 -m unittest tests.test_unit
